feat(user): make labels an option
Deprecate several methods that assume the field always exists.

Create several methods that replace the old methods but return an option
instead.

This change is motivated by the fact that there are now data sources
other than Warpcast Spam Labels and this struct is useful even without
the labels field.

Also make changes to the app and cli to remove deprecated methods