feat!(InvalidInputError): remove
This commit replaces InvalidInputError with SpamDataParseError, a similar
error but with a better name.

It changes the associated Error in TryFrom<UnprocessedUserLine> for
Fidded<DatedSpamUpdate> to SpamDataParseError from InvalidInputError.

This commit creates SpamScoreError, an error used in SpamDataParseError.

It replaces InvalidInputError with SpamScoreError in TryFrom<usize> for
SpamScore, a more reasonable error for this implementation.

It also removes the use of InvalidInputError and introduces a very
general InvalidInputError to DataCreationError. DataCreationError does
not use SpamDataParseError, since this should be a more general Error
type.